+++ src/modem/freq.rs
@@ -3,12 +3,12 @@ use std;
 #[derive(Copy, Clone)]
 pub struct Freq {
     // Cycles per second.
-    hz: u32,
-    sr: u32,
+    hz: usize,
+    sr: usize,
 }
 
 impl Freq {
-    pub fn new(hz: u32, sr: u32) -> Freq {
+    pub fn new(hz: usize, sr: usize) -> Freq {
         Freq {
             hz: hz,
             sr: sr,
@@ -27,7 +27,7 @@ impl Freq {
 
     // Calculate the number of samples needed to go through the given number of
     // cycles of the wave.
-    pub fn samples_for_cycles(&self, cycles: u32) -> usize {
+    pub fn samples_for_cycles(&self, cycles: usize) -> usize {
         (self.sr as f64 / self.hz as f64 * cycles as f64) as usize
     }
 }
